+++ index.ts
@@ -1,5 +1,6 @@
 import * as ieee754 from 'ieee754'
 import * as net from 'net'
+var strLenIdent = numTo128
 export type Listener = (key: string, value: any, valueType: String, type: "add" | "delete" | "update" | "flagChange", id: number, flags: number) => any
 export class Client {
     serverName: String
@@ -14,7 +15,8 @@ export class Client {
     private listeners: Listener[] = []
     private RPCExecCallback: { [key: number]: (result: Object) => any } = {}
     private lateCallbacks: (() => any)[] = []
-    private conCallback: (connected: boolean, err: Error) => any
+    private conCallback: (connected: boolean, err: Error, is2_0: boolean) => any
+    private is2_0 = false
     /**
      * True if the Client has completed its hello and is connected
      */
@@ -22,12 +24,18 @@ export class Client {
         return this.connected
     }
     /**
+     * True if the client has switched to 2.0
+     */
+    uses2_0() {
+        return this.is2_0
+    }
+    /**
      * Start the Client
-     * @param callback Called When an error occurs
+     * @param callback Called on connect or error
      * @param address Address of the Server. Default = "localhost"
      * @param port Port of the Server. Default = 1735
      */
-    start(callback?: (connected: boolean, err: Error) => any, address = '127.0.0.1', port = 1735) {
+    start(callback?: (connected: boolean, err: Error, is2_0: boolean) => any, address = '127.0.0.1', port = 1735) {
         this.connected = false
         this.address = address
         this.port = port
@@ -42,7 +50,7 @@ export class Client {
             if (this.reconnect) {
                 this.start(callback, address, port)
             }
-        }).on('error', err => callback(false, err))
+        }).on('error', err => callback(false, err, this.is2_0))
     }
     /**
      * Add a Listener to be called on change of an Entry
@@ -90,13 +98,21 @@ export class Client {
         /** Protocol Version Unsupported */
         0x02: (buf, off) => {
             var ver = `${buf[off++]}.${buf[off++]}`
-            if (ver === '2.0') this.reconnect = true
-            else this.conCallback(false, new Error('Unsupported protocol: ' + ver))
+            if (ver === '2.0') {
+                this.reconnect = true
+                this.is2_0 = true
+                strLenIdent = numTo2Byte
+            }
+            else this.conCallback(false, new Error('Unsupported protocol: ' + ver), this.is2_0)
             return off
         },
         /** Server Hello Complete */
         0x03: (buf, off) => {
-            this.toServer.HelloComplete()
+            if (this.is2_0) {
+                this.afterConnect()
+            } else {
+                this.toServer.HelloComplete()
+            }
             return off
         },
         /** Server Hello */
@@ -227,37 +243,47 @@ export class Client {
             return off
         }
     }
+    private afterConnect() {
+        this.connected = true
+        this.conCallback(true, null, this.is2_0)
+        while (this.lateCallbacks.length) {
+            this.lateCallbacks.shift()()
+        }
+    }
     private readonly toServer = {
         Hello: (serverName: string) => {
-            let s = TypeBuf[e.String].toBuf(serverName),
-                buf = Buffer.allocUnsafe(s.length + 3)
-            buf[0] = 0x01
-            buf[1] = 3
-            buf[2] = 0
-            s.write(buf, 3)
-            this.write(buf, true)
+            if (this.is2_0) {
+                this.write(toServer.hello2_0)
+            } else {
+                let s = TypeBuf[e.String].toBuf(serverName),
+                    buf = Buffer.allocUnsafe(s.length + 3)
+                buf[0] = 0x01
+                buf[1] = 3
+                buf[2] = 0
+                s.write(buf, 3)
+                this.write(buf, true)
+            }
         },
         HelloComplete: () => {
             this.write(toServer.helloComplete, true)
-            this.connected = true
-            this.conCallback(true, null)
-            while (this.lateCallbacks.length) {
-                this.lateCallbacks.shift()()
-            }
+            this.afterConnect()
         }
     }
     /**
      * Add an Entry
-     * @param type ID of the type of the Value
      * @param val The Value
      * @param name The Key of the Entry
      * @param persist Whether the Value should persist on the server through a restart
      */
-    Assign(type: number, val: any, name: string, persist = false) {
-        let n = TypeBuf[e.String].toBuf(name)
-        let f = TypeBuf[type].toBuf(val),
+    Assign(val: any, name: string, persist = false) {
+        let type = getType(val)
+        if (this.is2_0 && type === e.RawData) return new Error('2.0 does not have Raw Data')
+        if (type === e.RPC) return new Error('Clients can not assign an RPC')
+        let n = TypeBuf[e.String].toBuf(name),
+            f = TypeBuf[type].toBuf(val),
             nlen = n.length,
-            len = f.length + nlen + 7,
+            assignLen = this.is2_0 ? 6 : 7,
+            len = f.length + nlen + assignLen,
             buf = Buffer.allocUnsafe(len)
         buf[0] = 0x10
         n.write(buf, 1)
@@ -266,8 +292,8 @@ export class Client {
         buf[nlen + 3] = 0xff
         buf[nlen + 4] = 0
         buf[nlen + 5] = 0
-        buf[nlen + 6] = persist ? 1 : 0
-        f.write(buf, nlen + 7)
+        if (!this.is2_0) buf[nlen + 6] = persist ? 1 : 0
+        f.write(buf, nlen + assignLen)
         this.write(buf)
     }
     /**
@@ -281,7 +307,8 @@ export class Client {
         if (!checkType(val, entry.typeID)) return new Error('Wrong Type')
         entry.val = val
         let f = TypeBuf[entry.typeID].toBuf(val),
-            len = f.length + 6,
+            updateLen = this.is2_0 ? 5 : 6,
+            len = f.length + updateLen,
             buf = Buffer.allocUnsafe(len)
         entry.sn++
         buf[0] = 0x11
@@ -289,8 +316,8 @@ export class Client {
         buf[2] = id & 0xff
         buf[3] = entry.sn >> 8
         buf[4] = entry.sn & 0xff
-        buf[5] = entry.typeID
-        f.write(buf, 6)
+        if (!this.is2_0) buf[5] = entry.typeID
+        f.write(buf, updateLen)
         this.write(buf)
     }
     /**
@@ -299,6 +326,7 @@ export class Client {
      * @param persist Whether the Entry should persist through a restart on the server
      */
     Flag(id: number, persist = false) {
+        if (this.is2_0) return new Error('2.0 does not support flags')
         if (!(id in this.entries)) return new Error('Does not exist')
         this.write(Buffer.from([0x12, id >> 8, id & 0xff, persist ? 1 : 0]))
     }
@@ -307,6 +335,7 @@ export class Client {
      * @param id The ID of the Entry
      */
     Delete(id: number) {
+        if (this.is2_0) return new Error('2.0 does not support delete')
         if (!(id in this.entries)) return new Error('Does not exist')
         this.write(Buffer.from([0x13, id >> 8, id & 0xff]))
     }
@@ -314,6 +343,7 @@ export class Client {
      * Deletes All Entries
      */
     DeleteAll() {
+        if (this.is2_0) return new Error('2.0 does not support delete')
         this.write(toServer.deleteAll)
         this.entries = {}
         this.keymap = {}
@@ -325,6 +355,7 @@ export class Client {
      * @param callback To be called with the Results
      */
     RPCExec(id: number, val: Object, callback: (result: Object) => any) {
+        if (this.is2_0) return new Error('2.0 does not support RPC')
         if (id in this.entries) return new Error('Does not exist')
         let entry = this.entries[id]
         if (entry.typeID !== e.RPC) return new Error('Is not an RPC')
@@ -341,7 +372,7 @@ export class Client {
             len += n.length
             f.push(n)
         }
-        let encLen = len.to128(),
+        let encLen = numTo128(len),
             buf = Buffer.allocUnsafe(len + encLen.length + 5),
             off = 5 + encLen.length,
             randId = Math.floor(Math.random() * 0xffff)
@@ -403,9 +434,23 @@ function checkType(val: any, type: number) {
         else return false
     }
 }
+function getType(val: any) {
+    if (Array.isArray(val)) {
+        if (typeof val[0] === "boolean") return 0x10
+        else if (typeof val[0] === "number") return 0x11
+        else if (typeof val[0] === "string") return 0x12
+        else if (typeof val[0] === "object") return 0x20
+    } else {
+        if (typeof val === "boolean") return 0x00
+        else if (typeof val === "number") return 0x01
+        else if (typeof val === "string") return 0x02
+        else if (Buffer.isBuffer(val)) return 0x03
+    }
+}
 const toServer = {
     helloComplete: Buffer.from([0x05]),
-    deleteAll: Buffer.from([0xD0, 0x6C, 0xB2, 0x7A])
+    deleteAll: Buffer.from([0x14, 0xD0, 0x6C, 0xB2, 0x7A]),
+    hello2_0: Buffer.from([0x01, 2, 0])
 }
 export interface Entry {
     typeID: number,
@@ -502,7 +547,7 @@ const TypeBuf: fromBuf = {
     },
     0x02: <f<string>>{
         toBuf: (val) => {
-            let bufT = Buffer.concat([val.length.to128(), Buffer.from(val, 'utf8')])
+            let bufT = Buffer.concat([strLenIdent(val.length), Buffer.from(val, 'utf8')])
             return {
                 length: bufT.length,
                 write: (buf, off) => {
@@ -516,7 +561,7 @@ const TypeBuf: fromBuf = {
     },
     0x03: <f<Buffer>>{
         toBuf: (val) => {
-            let len = val.length.to128()
+            let len = numTo128(val.length)
             return {
                 length: val.length + len.length,
                 write: (buf, off) => {
@@ -592,7 +637,7 @@ const TypeBuf: fromBuf = {
             let lens: Buffer[] = [],
                 len = 1
             for (let i = 0; i < val.length; i++) {
-                lens[i] = Buffer.concat([val[i].length.to128(), Buffer.from(val[i])])
+                lens[i] = Buffer.concat([strLenIdent(val[i].length), Buffer.from(val[i])])
                 len += lens[i].length
             }
             return {
@@ -695,13 +740,15 @@ var TypesFrom: typesFrom = {
     0x20: TypeBuf[e.RPC].fromBuf,
     //0x21: TypeBuf[e.Byte].fromBuf
 }
-function fromLEBuf(buf, offset) {
+function fromLEBuf(buf: Buffer, offset: number) {
     let res = numFrom128(buf, offset),
         end = res.offset + res.val
     return { offset: end, val: buf.slice(res.offset, end).toString('utf8') }
 }
-Number.prototype.to128 = function (this: number) {
-    let n = this
+
+
+function numTo128(num: number) {
+    let n = num
     let r: number[] = []
     while (n > 0x07f) {
         r.push((n & 0x7f) | 0x80)
@@ -709,6 +756,9 @@ Number.prototype.to128 = function (this: number) {
     }
     r.push(n)
     return Buffer.from(r)
+}
+function numTo2Byte(num: number) {
+    return Buffer.from([(this >> 8) & 0xff, this & 0xff])
 }
 function numFrom128(buf: Buffer, offset: number) {
     let r = 0, n = buf[offset]
@@ -722,13 +772,5 @@ function numFrom128(buf: Buffer, offset: number) {
     return {
         val: r,
         offset
-    }
-}
-declare global {
-    interface Number {
-        /**
-         * Converts a number to a Buffer using LEB128
-         */
-        to128(): Buffer
     }
 }
